### 2024-02-10 #6343 @editorialbot
[PRE REVIEW]: Anime: Atmospheric and Instrument Models in the Measurement Equation
**Submitting author:** @iniyannatarajan (<a href="http://orcid.org/0000-0001-8242-4373">Iniyan Natarajan</a>)
**Repository:** https://github.com/iniyannatarajan/Anime.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.0
