+++ src/plots.jl
@@ -1,31 +1,4 @@
-export plotuvcov, plotvis, plotstationgains, plotbandpass, plotpointingerrors, plotelevationangle, plotparallacticangle, plotdterms, plottransmission,
-plotmeandelays
-
-using Plots.PlotMeasures
-
-"""
-    plotuvcov(uvw::Matrix{Float64}, flagrow::Vector{Bool}, chanfreqvec::Vector{Float64}; saveprefix="test_")
-
-Plot uv-coverage of observation.
-"""
-function plotuvcov(uvw::Matrix{Float64}, flagrow::Vector{Bool}, chanfreqvec::Vector{Float64}; saveprefix="test_")
-    @info("Generating uv-coverage plots...")
-
-    maskindices = findall(isequal(true), flagrow)
-    muwave = deepcopy(uvw[1,:]) / (299792458.0/mean(chanfreqvec)) / 1e9 # convert to Gλ
-    mvwave = deepcopy(uvw[2,:]) / (299792458.0/mean(chanfreqvec)) / 1e9
-
-    muwave[maskindices] .= NaN
-    mvwave[maskindices] .= NaN
-
-    p = plot()
-    plot!(p, size=(1050, 700), guidefontsize=18, legendfontsize=18, left_margin=5mm, bottom_margin=10mm)
-    plot!(p, muwave, mvwave, seriestype=:scatter, ls=:dot, ms=1, mc=:blue, msc=:blue, label="", xflip=true)
-    plot!(p, xlabel="u (Gλ)", ylabel="v (Gλ)", title="uv-coverage")
-
-    savefig(p, saveprefix*"uvcoverage.png")
-    @info("Done 🙆")
-end
+export plotvis, plotstationgains
 
 """
     plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Array{Bool,4}, data::Array{Complex{Float32},4},
@@ -42,101 +15,51 @@ function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Arra
     maskeddata = deepcopy(data)
     maskeddata[maskindices] .= NaN # assign NaN values to flagged visibilities
 
+    #set pol colours
+    colours = [[:red, :cyan], [:purple, :green]]
+    labels = [["RR", "RL"], ["LR", "LL"]]
+
     # plot visibility amplitudes against projected baseline separation
-    p = plot()
-    plot!(p, size=(1050, 700), guidefontsize=18, legendfontsize=18, left_margin=5mm, bottom_margin=10mm)
-    # plot first frequency channel
-    plot!(p, uvwave, abs.(maskeddata[1,1,1,:]), seriestype=:scatter, ls=:dot, ms=1, mc=:red, msc=:red, label="RR")
-    plot!(p, uvwave, abs.(maskeddata[1,2,1,:]), seriestype=:scatter, ls=:dot, ms=1, mc=:cyan, msc=:cyan, label="RL")
-    plot!(p, uvwave, abs.(maskeddata[2,1,1,:]), seriestype=:scatter, ls=:dot, ms=1, mc=:purple, msc=:purple, label="LR")
-    plot!(p, uvwave, abs.(maskeddata[2,2,1,:]), seriestype=:scatter, ls=:dot, ms=1, mc=:green, msc=:green, label="LL")
-    
-    # plot the rest of the frequency channels
-    if numchan > 1
-        plot!(p, uvwave, abs.(maskeddata[1,1,2:end,:]'), seriestype=:scatter, ls=:dot, ms=1, mc=:red, msc=:red, label="")
-        plot!(p, uvwave, abs.(maskeddata[1,2,2:end,:]'), seriestype=:scatter, ls=:dot, ms=1, mc=:cyan, msc=:cyan, label="")
[…]
