+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.3.0"
 ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
 CFITSIO_jll = "b3e40c51-02ae-5482-8a39-3ace5868dcf4"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 Casacore = "72bb1062-94f5-49fa-bb69-94b615203ad9"
 ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
 CondaPkg = "992eb4ea-22a4-4c89-a5bb-47a3300528ab"
@@ -24,7 +25,6 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 OpenSSL_jll = "458c3c95-2e84-50aa-8efc-19380b2a3a95"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
@@ -37,6 +37,7 @@ YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 ArgParse = "1.1"
 CFITSIO_jll = "=4.0.0"
 CSV = "0.10"
+CairoMakie = "0.11"
 Casacore = "0.2"
 ColorSchemes = "3.21"
 CondaPkg = "0.2"
@@ -52,7 +53,6 @@ LaTeXStrings = "1.3.0"
 LazyGrids = "0.5"
 Literate = "2.14"
 OpenSSL_jll = "=1.1.21"
-Plots = "1.38"
 PythonCall = "0.9"
 SpecialFunctions = "2.3"
 Statistics = "1"
