### 2024-02-20 22a2b14
Update plotting functions (WIP)
