+++ src/plots.jl
@@ -1,4 +1,4 @@
-export plotvis, plotstationgains
+export plotvis, plotstationgains, plotbandpass
 
 """
     plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Array{Bool,4}, data::Array{Complex{Float32},4},
@@ -7,13 +7,15 @@ export plotvis, plotstationgains
 Plot complex visibilities against time and projected baseline length.
 """
 function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Array{Bool,4}, data::Array{Complex{Float32},4},
-    numchan::Int64, times::Vector{Float64}; plotphases::Bool=false, saveprefix="data_")
+    numchan::Int64, times::Vector{Float64}; plotphases::Bool=false, saveprefix="data")
     @info("Generating visibility plots...")
     uvwave = sqrt.(uvw[1,:].^2 .+ uvw[2,:].^2) / (299792458.0/mean(chanfreqvec)) / 1e9 # in units of Gλ
 
     maskindices = findall(isequal(true), flag)
     maskeddata = deepcopy(data)
     maskeddata[maskindices] .= NaN # assign NaN values to flagged visibilities
+
+    xtimes = times .- times[1] # relative time in seconds
 
     #set pol colours
     colours = [[:red, :cyan], [:purple, :green]]
@@ -24,15 +26,17 @@ function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Arra
     ax = Axis(f[1, 1], xlabel="Projected baseline separation (Gλ)", ylabel="Complex visibility amplitude (Jy)")
     for ii in 1:2
         for jj in 1:2
-            scatter!(ax, uvwave, abs.(maskeddata[ii,jj,1,:]), color=colours[ii][jj], label=labels[ii][jj], markersize=1)
+            scatter!(ax, uvwave, abs.(maskeddata[ii,jj,1,:]), color=colours[ii][jj], label=labels[ii][jj], markersize=2)
             if numchan > 1
-                scatter!(ax, uvwave, abs.(maskeddata[ii,jj,2:end,:]'), color=colours[ii][jj], label="", markersize=1)
+                for channo in 2:numchan
+                    scatter!(ax, uvwave, abs.(maskeddata[ii,jj,channo,:]), color=colours[ii][jj], markersize=2)
+                end
             end
         end
     end
 
-    axislegend(ax, position=:rt)
-    save(saveprefix*"visibilityamplitude_vs_baseline.png", f)
+    axislegend(ax, position=:rt, merge=true, unique=true, tellheight=true, tellwidth=true)
+    save(saveprefix*"_visibilityamplitude_vs_baseline.png", f)
 
     # plot visibility phases against projected baseline separation
     if plotphases
@@ -40,15 +44,53 @@ function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Arra
         ax = Axis(f[1, 1], xlabel="Projected baseline separation (Gλ)", ylabel="Complex visibility phase (deg.)")
         for ii in 1:2
             for jj in 1:2
-                scatter!(ax, uvwave, rad2deg.(angle.(maskeddata[ii,jj,1,:])), color=colours[ii][jj], label=labels[ii][jj], markersize=1)
+                scatter!(ax, uvwave, rad2deg.(angle.(maskeddata[ii,jj,1,:])), color=colours[ii][jj], label=labels[ii][jj], markersize=2)
                 if numchan > 1
[…]
