### 2024-02-20 828fb24
Update plotting functions (WIP)
