+++ src/plots.jl
@@ -1,4 +1,4 @@
-export plotvis, plotstationgains, plotbandpass
+export plotvis, plotstationgains, plotbandpass, plotpointingerrors
 
 """
     plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Array{Bool,4}, data::Array{Complex{Float32},4},
@@ -22,7 +22,7 @@ function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Arra
     labels = [["RR", "RL"], ["LR", "LL"]]
 
     # plot visibility amplitudes against projected baseline separation
-    f = Figure(size=(900, 600))
+    f = Figure(size=(800, 400))
     ax = Axis(f[1, 1], xlabel="Projected baseline separation (Gλ)", ylabel="Complex visibility amplitude (Jy)")
     for ii in 1:2
         for jj in 1:2
@@ -40,7 +40,7 @@ function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Arra
 
     # plot visibility phases against projected baseline separation
     if plotphases
-        f = Figure(size=(900, 600))
+        f = Figure(size=(800, 400))
         ax = Axis(f[1, 1], xlabel="Projected baseline separation (Gλ)", ylabel="Complex visibility phase (deg.)")
         for ii in 1:2
             for jj in 1:2
@@ -58,7 +58,7 @@ function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Arra
     end
 
     # plot visibility amplitudes against time
-    f = Figure(size=(900, 600))
+    f = Figure(size=(800, 400))
     ax = Axis(f[1, 1], xlabel="Projected baseline separation (Gλ)", ylabel="Complex visibility amplitude (Jy)")
     for ii in 1:2
         for jj in 1:2
@@ -76,7 +76,7 @@ function plotvis(uvw::Matrix{Float64}, chanfreqvec::Array{Float64,1}, flag::Arra
 
     # plot visibility phases against time
     if plotphases
-        f = Figure(size=(900, 600))
+        f = Figure(size=(800, 400))
         ax = Axis(f[1, 1], xlabel="Projected baseline separation (Gλ)", ylabel="Complex visibility phase (deg.)")
         for ii in 1:2
             for jj in 1:2
@@ -111,7 +111,7 @@ function plotstationgains(h5file::String, scanno::Vector{Int32}, times::Vector{F
     # get unique times
     uniqtimes = unique(times)
 
-    f = Figure(size=(900, 400))
+    f = Figure(size=(500, 300))
     axphase1 = Axis(f[1, 1], ylabel="Phases (°)", title="Station gains (Pol1)")
     axamp1 = Axis(f[2, 1], xlabel="Relative time (hr)", ylabel="Amplitudes")
     axphase2 = Axis(f[1, 2], title="Station gains (Pol2)")
@@ -166,7 +166,6 @@ function plotstationgains(h5file::String, scanno::Vector{Int32}, times::Vector{F
     linkyaxes!(axphase1, axphase2)
 
     f[1:2, 3] = Legend(f, axamp1, merge=true, unique=true, tellheight=true, tellwidth=true)
-
     save("StationGains_vs_time.png", f)
 
     @info("Plotted station gains 🙆")
@@ -185,10 +184,10 @@ function plotbandpass(h5file::String, stationnames::Vector{String3}, chanfreqvec
 
     b = read(fid["bandpass"]["bjonesmatrices"])
 
-    f = Figure(size=(900, 400))
[…]
