### 2024-02-20 870005e
Plot elevation angles
