### 2024-02-20 5a8ce49
Plot parallactic angle
