### 2024-02-20 5f9377b
Plot D-terms
