### 2024-02-20 b495783
Plot mean delays
