### 2024-02-20 4fa3a03
Plot uv-coverage
