### 2024-02-20 e668018
Use scatter plot
