### 2024-02-21 7ae193f
Display plots in docs
