+++ .github/workflows/CompatHelper.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Install Julia, but only if it is not already available in the PATH
         uses: julia-actions/setup-julia@v1
         with:
-          version: '1.9'
+          version: '1.10'
           arch: ${{ runner.arch }}
         if: steps.julia_in_path.outcome != 'success'
       - name: "Add the General registry via Git"
