+++ .github/workflows/ci.yml
@@ -26,6 +26,7 @@ jobs:
       matrix:
         version:
           - '1.10'
+          - '1'
         os:
           - ubuntu-latest
         arch:
