+++ .github/workflows/ci.yml
@@ -25,8 +25,8 @@ jobs:
     strategy:
       matrix:
         version:
+          - '1.9'
           - '1.10'
-          - '1'
         os:
           - ubuntu-latest
         arch:
