### 2024-03-06 c2ba6ce
Restrict Anime 0.4 to to Julia version 1.10 or above
