+++ Project.toml
@@ -59,7 +59,7 @@ Statistics = "1"
 StatsBase = "0.33, 0.34"
 Tables = "1.10"
 YAML = "0.4"
-julia = "=1.9, =1.10"
+julia = "1.10"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
