+++ Project.toml
@@ -53,7 +53,7 @@ LaTeXStrings = "1.3.0"
 LazyGrids = "0.5"
 Literate = "2.14"
 OpenSSL_jll = "=1.1.21"
-PythonCall = "0.9"
+PythonCall = "=0.9.15"
 SpecialFunctions = "2.3"
 Statistics = "1"
 StatsBase = "0.33, 0.34"
