### 2024-04-02 #6573 @editorialbot
[REVIEW]: Anime: Atmospheric and Instrument Models in the Measurement Equation
**Submitting author:** @iniyannatarajan (<a href="http://orcid.org/0000-0001-8242-4373">Iniyan Natarajan</a>)
**Repository:** https://github.com/iniyannatarajan/Anime.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.0
