### 2024-04-02 iniyannatarajan/Anime.jl#26 @michael-petersen
macOS install?
openjournals/joss-reviews#6573

Hi, is it possible to get a working install recipe for macOS? Currently the Julia Pkg install fails out of the box on an M1 machine.

I noticed that macOS-latest is excluded from the OS matrix, https://github.com/iniyannatarajan/Anime.jl/blob/9f9423f955c18c525078c832a6f984e127777116/.github/workflows/ci.yml#L37C17-L37C29, and the documentation notes that Anime has been tested on Ubuntu Linux. It seem that `casatools` may have some issues with M1/2/3 machines, but that they also may be surmountable.
