### 2024-04-08 iniyannatarajan/Anime.jl#26 @iniyannatarajan
Hi, sorry for the delay. I do not have access to a Mac, but I have put together the following ci.yml based on what I can find online for Mac support:

```
name: CI

# Controls when the workflow will run
on:
  # Triggers the workflow on push or pull request events but only for the "main" branch
  push:
    branches: [ "main" ]
  pull_request:
    branches: [ "main" ]

  # Allows you to run this workflow manually from the Actions tab
  workflow_dispatch:

# A workflow run is made up of one or more jobs that can run sequentially or in parallel
jobs:
  # This workflow contains two jobs called "build" and "docs"
  build:
    name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
    # defines the type of runner that the job will run on
    runs-on: ${{ matrix.os }}

    # defines matrix with a bunch of options
    strategy:
      matrix:
        version:
          - '1.10'
          - '1'
        os:
          - ubuntu-latest
          - macOS-latest
        arch:
          - x64
        python-version:
          - '3.8'
        exclude:
          - os: macOS-latest
          - arch: x86

    # Steps represent a sequence of tasks that will be executed as part of the job
    steps:
      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
      - uses: actions/checkout@v3

      # Sets up python environment
      - name: Setup python ${{ matrix.python-version }}
        uses: actions/setup-python@v4
        with:
          python-version: ${{ matrix.python-version }}
          architecture: ${{ matrix.arch }}

      # Sets up venv and installs python dependencies for macOS
      - name: Setup venv and install dependencies (macOS)
        if: runner.os == 'macOS'
        run: |
          arch -x86_64 python3.10 -m venv myvenv
          source myvenv/bin/activate
          arch -x86_64 python3 -m pip install --upgrade pip setuptools wheel casatools

      # Installs python dependencies for other OS
      - name: Install python dependencies
        if: runner.os != 'macOS'
        run: |
          python -m pip install --upgrade pip setuptools wheel
          python -m pip install casatools

      # Sets up Julia environment
      - name: Setup Julia ${{ matrix.version }}
        uses: julia-actions/setup-julia@v1
        with:
          version: ${{ matrix.version }}
          arch: ${{ matrix.arch }}
      - uses: julia-actions/julia-buildpkg@v1
        env:
          PYTHON: python

      # Run tests
      - name: Run tests
        uses: julia-actions/julia-runtest@v1

      - uses: julia-actions/julia-processcoverage@v1

      - name: Upload coverage reports to Codecov
        uses: codecov/codecov-action@v3
        env:
          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
```

I have modified the casatools installation instructions in the workflow file per [this link](https://casadocs.readthedocs.io/en/stable/notebooks/introduction.html#Modular-Packages).

Please let me know if this works for you. Otherwise I can add some pointers for Mac users in the official documentation.

Thanks!
