### 2024-04-15 #6573 @iniyannatarajan
Hi Kiran,

Thanks very much for your comments! I have added a Contributing section to the README.

I see this error with the latest PythonCall. PythonCall has had a few [releases](https://github.com/JuliaPy/PythonCall.jl/releases) since v0.9.15, which trip up the arguments being passed to the Python functions by msfromconfig function. Restricting PythonCall to v0.9.15 gets rid of this error.

The version tagged to the REVIEW is 0.3.0 but I have made a 0.4.0 release since then. The PyhtonCall releases happened after I released this version, so I have updated the Project.toml on the main repo restricting PythonCall to v0.9.15. So, the latest version on the main branch should not give this error.

Would it be possible for you (and would it be okay) to test the main branch for the review, or should I make a new release with only this change made to Project.toml?

Also, I will expand the CI to other platforms, but I do not have ready access to them. Any idea on how to easily debug if I run into error on the other platforms? Thanks!
