+++ test/Project.toml
@@ -15,7 +15,6 @@ Interpolations = "a98d9a8b-a2ab-59e6-89dd-64a1c18fca59"
 LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 OpenSSL_jll = "458c3c95-2e84-50aa-8efc-19380b2a3a95"
 PythonCall = "6099a3de-0909-46bc-b1f4-468b9a2dfc0d"
