### 2024-04-17 0981894
Update docs with how to run tutorial scripts
