+++ docs/src/index.md
@@ -12,6 +12,18 @@ Within the EHT, `MeqSilhouette v2 (MEQSv2)`[^IN2022][^TB2017] is used to generat
 
 `Anime` is designed to provide a complete framework in Julia to compute instrument models from first principles wherever possible and to generate synthetic VLBI data sets with support for and convert b/w multiple VLBI data storage formats. It can construct instrument models for multi-scan VLBI observations with irregulary-spaced and missing data without having to construct a regular grid of complex visibilities in baseline and time (like `SYMBA/MEQSv2`), thereby speeding up the generation of synthetic data. It also constructs more realistic models, re-implementing existing models from `SYMBA/MEQSv2` in `Julia` for more efficient computation and introducing new ones. The ultimate aim is to be able to read/write commonly used on-disk data formats in VLBI and provide a set of instrument modelling functions that other software written in `Julia` (e.g. `Comrade`[^PT2022]) can import and use for synthetic data generation, calibration and imaging. Inference methods for emulating the statistical properties of actual data are also in development.
 
+The documentation is structured as follows:
+- *Installation* describes the installation procedure for `Anime` and optional external software necessary to perform full synthetic data generation as opposed to only generating instrument models.
+- *Components* describes the various input files and their formats necessary to run the different parts of `Anime`.
+- *Instrument Models* describes some theory behind the models that are currently supported.
+- *Tutorial* consists of a set of scripts that cover all the basic functionality of `Anime`. The entire documentation, including the outputs for the tutorial scripts can be generated by cloning the repo from GitHub, changing into the *examples/* directory and typing the following:
+```julia
+$ julia --project=docs/
+(docs) pkg> instantiate
+(docs) pkg> dev .
+julia> include("docs/make.jl")
+```
+
 ```@contents
 Pages = [
     "index.md",
