### 2024-04-17 3cf4e17
Minor update to docs for generating tutorials locally
