+++ docs/src/index.md
@@ -16,8 +16,10 @@ The documentation is structured as follows:
 - *Installation* describes the installation procedure for `Anime` and optional external software necessary to perform full synthetic data generation as opposed to only generating instrument models.
 - *Components* describes the various input files and their formats necessary to run the different parts of `Anime`.
 - *Instrument Models* describes some theory behind the models that are currently supported.
-- *Tutorial* consists of a set of scripts that cover all the basic functionality of `Anime`. The entire documentation, including the outputs for the tutorial scripts can be generated by cloning the repo from GitHub, changing into the *examples/* directory and typing the following:
+- *Tutorial* consists of a set of scripts that cover all the basic functionality of `Anime`. The entire documentation, including the outputs for the tutorial scripts can be generated by typing the following:
 ```julia
+$ git clone https://github.com/iniyannatarajan/Anime.jl.git
+$ cd Anime.jl
 $ julia --project=docs/
 (docs) pkg> instantiate
 (docs) pkg> dev .
