+++ src/storageutils/loadobs.jl
@@ -1,230 +1,0 @@
-export loadms, CjlObservation
-
-abstract type AbstractObservation{T} end
-
-"""
-    $(TYPEDEF)
-
-Container type for storing observation parameters and data.
-
-# Fields
-$(FIELDS)
-"""
-struct CjlObservation{T} <: AbstractObservation{T}
-    """
-    Name of the Measurement Set
-    """
-    msname::String
-    """
-    Complex visibility data (MS `DATA` column)
-    """
-    data::Array{Complex{Float32},4}
-    """
-    Flag array of the same dimensions as data (MS `FLAG` column)
-    """
-    flag::Array{Bool,4}
-    """
-    Flag row Boolean vector (MS ```FLAG_ROW``` column)
-    """
-    flagrow::Array{Bool,1}
-    """
-    Antenna 1 in a baseline pair (MS `ANTENNA1` column)
-    """
-    antenna1::Vector{Int32}
-    """
-    Antenna 2 in a baseline pair (MS `ANTENNA2` column)
-    """
-    antenna2::Vector{Int32}
-    """
-    uvw coordinates (MS `UVW` column)
-    """
-    uvw::Matrix{Float64}
-    """
-    Timestamps (MS `TIME` column)
-    """
-    times::Vector{Float64}
-    """
-    Integration time
-    """
-    exposure::Float64
-    """
-    Scan numbers (MS ```SCAN_NUMBER``` column)
-    """
-    scanno::Vector{Int32}
-    #=weight::Vector{Vector{Float32}}
-    weightspec::Array{Float32,3}
-    sigma::Vector{Vector{Float32}}
-    sigmaspec::Array{Float32,3}=#
-    """
-    Number of frequency channels
-    """
-    numchan::Int64
-    """
-    Channel frequencies (Hz) (MS ```CHAN_FREQ``` column)
-    """
-    chanfreqvec::Array{Float64,1}
-    """
-    Width of frequency channel
-    """
-    chanwidth::Float64
-    """
-    Direction of phase centre
-    """
-    phasedir::Array{Float64,2}
-    """
-    Antenna positions in x,y,z coordinates
-    """
-    pos::Array{Float64, 2}
-    """
-    Dataframe of station information input by user
-    """
-    stationinfo::DataFrame
-    #yamlconf::Dict
-    """
-    Consider only water vapour in the troposphere
-    """
-    tropwetonly::Bool
-    """
-    Correlator efficiency
-    """
-    correff::Float64
-    """
-    Introduce attenuation by the troposphere
-    """
-    tropattenuate::Bool
-    """
-    Introduce noise due to troposphere
-    """
-    tropskynoise::Bool
-    """
-    Introduce mean delays due to troposphere
-    """
-    tropmeandelays::Bool
-    """
-    Introduce turbulence in troposphere
-    """
-    tropturbulence::Bool
-    """
-    Polarization frame
-    """
-    polframe::String
-    """
-    Mode to use to create frequency-varying D-term samples
-    """
-    polmode::String
-    """
-    Pointing interval (s)
-    """
-    ptginterval::Float64
-    """
-    Pointing scale mixture parameter
-    """
-    ptgscale::Float64
-    """
-    Mode to use to create pointing error time samples
-    """
-    ptgmode::String
-    """
-    Mode to use to create station gain time samples
-    """
-    stationgainsmode::String
-    """
-    Input bandpass file
-    """
[…]

[…]
