+++ src/io/read.jl
@@ -1,0 +1,225 @@
+export readms, readobsconfig, readstationinfo, readbandpassinfo, readalistv5, readalistv6, MeasurementSet
+
+"""
+    $(TYPEDEF)
+
+Composite type for storing data from a Measurement Set.
+
+# Fields
+$(FIELDS)
+"""
+struct MeasurementSet{T}
+    """
+    Measurement Set name
+    """
+    name::String
+    """
+    Complex visibility data (MS `DATA` column)
+    """
+    data::Array{Complex{Float32},4}
+    """
+    Flag array of the same dimensions as data (MS `FLAG` column)
+    """
+    flag::Array{Bool,4}
+    """
+    Flag row Boolean vector (MS ```FLAG_ROW``` column)
+    """
+    flagrow::Array{Bool,1}
+    """
+    Antenna 1 in a baseline pair (MS `ANTENNA1` column)
+    """
+    antenna1::Vector{Int32}
+    """
+    Antenna 2 in a baseline pair (MS `ANTENNA2` column)
+    """
+    antenna2::Vector{Int32}
+    """
+    uvw coordinates (MS `UVW` column)
+    """
+    uvw::Matrix{Float64}
+    """
+    Timestamps (MS `TIME` column)
+    """
+    times::Vector{Float64}
+    """
+    Integration time
+    """
+    exposure::Float64
+    """
+    Scan numbers (```SCAN_NUMBER``` column in MS)
+    """
+    scanno::Vector{Int32}
+    """
+    Number of frequency channels
+    """
+    numchan::Int64
+    """
+    Channel frequencies (Hz) (MS ```CHAN_FREQ``` column)
+    """
+    chanfreqvec::Array{Float64,1}
+    """
+    Width of frequency channel
+    """
+    chanwidth::Float64
+    """
+    Direction of phase centre
+    """
+    phasedir::Array{Float64,2}
+    """
+    Antenna positions in x,y,z coordinates
+    """
+    pos::Array{Float64, 2}
+end
+
+"""
+    readms(msname::String)
+
+Read data from Measurement Set.
+"""
+function readms(msname::String)
+    tab = CCTable(msname, CCTables.Old)
+
+    # read values from ms
+    data::Vector{Matrix{ComplexF32}} = tab[:DATA][:]
+    flag::Vector{Matrix{Bool}} = tab[:FLAG][:]
+    flagrow::Vector{Bool} = tab[:FLAG_ROW][:]
+    antenna1::Vector{Int32} = tab[:ANTENNA1][:]
+    antenna2::Vector{Int32} = tab[:ANTENNA2][:]
+    uvw::Matrix{Float64} = tab[:UVW][:,:]
+    times::Vector{Float64} = tab[:TIME][:]
+    exposure::Float64 = tab[:EXPOSURE][1]
+    scanno::Vector{Int32} = tab[:SCAN_NUMBER][:]
+    #=weight::Vector{Vector{Float32}} = tab[:WEIGHT][:]
+    if :WEIGHT_SPECTRUM in keys(tab)
+        weight_spectrum::Vector{Matrix{Float32}} = tab[:WEIGHT_SPECTRUM][:,:,:]
+    else
+        weight_spectrum = fill(fill(missing, (2, 2)), size(data))
+    end
+
+    sigma::Vector{Vector{Float32}} = tab[:SIGMA][:]
+    if :SIGMA_SPECTRUM in keys(tab)
+        sigma_spectrum::Vector{Matrix{Float32}} = tab[:SIGMA_SPECTRUM][:,:,:]
+    else
+        sigma_spectrum = fill(fill(missing, (2, 2)), size(data))
+    end=#
+
+    spectab = tab.SPECTRAL_WINDOW
+    numchan::Int32 = spectab[:NUM_CHAN][1]
[…]
