+++ src/io/write.jl
@@ -1,0 +1,365 @@
+export makecasaantennatable, addweightcolumns, createmsfromuvfits, createuvfitsfromms, createmsfromconfig, writems
+
+using StatsBase: mode
+
+"""
+    makecasaantennatable(df::DataFrame, casaanttemplate::String; delim::String=",", ignorerepeated::Bool=false)
+
+Create CASA antenna table from the DataFrame df containing station information using `casaanttemplate` as template. Requires `casatools`.
+"""
+function makecasaantennatable(df::DataFrame, casaanttemplate::String; delim::String=",", ignorerepeated::Bool=false)    
+    #stationtable = "ANTENNA_$(split(basename(stations), '.')[1])"
+    stationtable = "ANTENNA_temp"
+
+    # read template table and copy
+    tb.open(casaanttemplate)
+    tb.copy(stationtable, deep=true, valuecopy=true)
+    tb.close()
+    tb.clearlocks()
+
+    # populate new table with values from the DataFrame
+    tb.open(stationtable, nomodify=false)
+    for ii in 1:length(df.station)-1
+        tb.copyrows(stationtable, nrow=1)
+    end
+    for ii in 1:length(df.station)
+        tb.putcol("STATION", String(df.station[ii]), startrow=ii-1, nrow=1)
+        tb.putcol("NAME", String(df.station[ii]), startrow=ii-1, nrow=1)
+        tb.putcol("MOUNT", String(df.mount[ii]), startrow=ii-1, nrow=1)
+    end
+    tb.putcol("DISH_DIAMETER", PyList(df.dishdiameter_m))
+    #for jj in 1:length(df.station)
+    #	tb.putcell("POSITION", jj-1, PyList(parse.(Float64, split.(df.xyzpos_m,',')[jj])))
+    #end
+    index = 0
+    for (x,y,z) in zip(df.x_m,df.y_m,df.z_m)
+        tb.putcell("POSITION", index, PyList([x,y,z]))
+        index += 1
+    end
+    tb.close()
+    tb.clearlocks()
+
+    return stationtable
+end
+
+"""
+    addweightcolumns(msname::String, mode::String, sigmaspec::Bool, weightspec::Bool)
+
+Add ```WEIGHT_SPECTRUM``` and ```SIGMA_SPECTRUM``` columns to existing MS. Requires `casatools` to be installed.
+"""
+function addweightcolumns(msname::String, mode::String, sigmaspec::Bool, weightspec::Bool)
+    # get quantities to define array shapes
+    tb.open("$(msname)::SPECTRAL_WINDOW")
+    numchan = pyconvert(Int64, tb.getcol("NUM_CHAN")[0])
+    tb.close()
+    
+    tb.open("$(msname)::POLARIZATION")
+    numcorr = pyconvert(Int64, tb.getcol("NUM_CORR")[0])
+    tb.close()
+
+    shape = PyList([numchan, numcorr])
+
+    tb.open(msname, nomodify=false)
+    if sigmaspec
+        coldesc = pydict(Dict("valueType" => "float", "dataManagerType" => "TiledShapeStMan", "dataManagerGroup" => "TiledSigmaSpectrum", "option" => 0, "maxlen" => 0, "comment" => "Estimated rms noise for each data point", "ndim" => 2, "_c_order" => true, "keywords" => pydict(Dict()), "shape" => shape))
+        tabdesc = pydict(Dict("SIGMA_SPECTRUM" => coldesc))
+	tb.addcols(tabdesc)
+	sigsp = tb.getcol("SIGMA_SPECTRUM")
+	PyArray(sigsp)[:] .= 1.0
+	tb.putcol("SIGMA_SPECTRUM", sigsp)
+    end
+
+    if mode == "uvfits"
[…]
