+++ examples/computemodels.jl
@@ -11,54 +11,59 @@ using .Anime
 using HDF5
 using CairoMakie
 
-# We also load two instances of *observation* from existing data sets to illustrate the various models that are generated -- a single-channel data set
-# and a multi-frequency data set.
-msname = joinpath(relativepath, "test", "data", "ehtuvf.ms")
-stations = joinpath(relativepath, "inputs", "eht_2017.stations")
-corruptseed = 456
-tropseed = 54364
-tropwetonly =false
-correff = 0.88
-tropattenuate = true
-tropskynoise = true
-tropmeandelays = true
-tropturbulence = true
-polvisframe = "sky"
-polmode = "gp"
-ptginterval = 5.0
-ptgscale = 2.0
-ptgmode = "gp"
-gainsmode = "gp"
-bpfile = joinpath(relativepath, "test", "data", "eht_2017.bandpass")
-delim = ","
-ignorerepeated = false
+# We will use two data sets to illustrate instrument model generation -- a single-channel data set
+# and a multi-frequency data set. We first prepare config Dicts for use with the two data sets.
+obsconfig = Dict()
 
-obs1 = loadms(msname, stations, corruptseed, tropseed, tropwetonly, correff, tropattenuate, tropskynoise, tropmeandelays, tropturbulence, polvisframe,
-polmode, ptginterval, ptgscale, ptgmode, gainsmode, bpfile, delim=",", ignorerepeated=false)
+obsconfig["msname"] = joinpath(relativepath, "test", "data", "ehtuvf.ms")
+obsconfig["stations"] = joinpath(relativepath, "inputs", "eht_2017.stations")
+obsconfig["corruptseed"] = 456
+obsconfig["troposphere"] = Dict()
+obsconfig["troposphere"]["tropseed"] = 54364
+obsconfig["troposphere"]["wetonly"] = false
+obsconfig["correff"] = 0.88
+obsconfig["troposphere"]["attenuate"] = true
+obsconfig["troposphere"]["skynoise"] = true
+obsconfig["troposphere"]["meandelays"] = true
+obsconfig["troposphere"]["turbulence"] = true
+obsconfig["instrumentalpolarization"] = Dict()
+obsconfig["instrumentalpolarization"]["visibilityframe"] = "sky"
+obsconfig["instrumentalpolarization"]["mode"] = "gp"
+obsconfig["pointing"] = Dict()
+obsconfig["pointing"]["interval"] = 5.0
+obsconfig["pointing"]["scale"] = 2.0
+obsconfig["pointing"]["mode"] = "gp"
+obsconfig["stationgains"] = Dict()
+obsconfig["stationgains"]["mode"] = "gp"
+obsconfig["bandpass"] = Dict()
+obsconfig["bandpass"]["bandpassfile"] = joinpath(relativepath, "test", "data", "eht_2017.bandpass")
 
-# Now load the multi-frequency data set
+singlechannelms = readms(joinpath(relativepath, "test", "data", "ehtuvf.ms")) # single-channel MS
+multichannelms = readms(joinpath(relativepath, "test", "data", "eht1.ms")) # multi-frequency MS
 
-obs2 = loadms(joinpath(relativepath, "test", "data", "eht1.ms"), stations, corruptseed, tropseed, tropwetonly, correff, tropattenuate, tropskynoise, tropmeandelays, tropturbulence, polvisframe,
-polmode, ptginterval, ptgscale, ptgmode, gainsmode, bpfile, delim=",", ignorerepeated=false)
+stationinfo = readstationinfo(obsconfig["stations"], delim=",", ignorerepeated=false) # read station info file
[…]
