### 2024-04-19 6d72cbc
Add tutorial for synthetic data
