### 2024-04-26 iniyannatarajan/Anime.jl#26 @iniyannatarajan
Hi, I was just wondering if the above CI file was helpful. Please let me know if you have any difficulties installing or running the package.

Thanks!
