### 2024-05-01 iniyannatarajan/Anime.jl#26 @lindyblackburn
I tried installing under the macOS x86 (Rosetta) version of Julia v1.10, but the CondaPkg pip environment only supports cp38-cp38-macosx_10_16 and earlier according to pip debug, while the pinned [casatools 6.5.5.21](https://pypi.org/project/casatools/6.5.5.21/#files) is only available for macosx_11_0.

I am able to directly pip install the following on M1 mac under the Julia x86 Conda environment,
casatools-6.5.1.23 numpy-1.24.4
