### 2024-05-06 #6573 @kiranshila
> Hi Kiran,
>
> I was wondering if you were able to get rid of the PythonCall error with the latest code in the main branch? I have also added instructions to generate all the tutorials locally in the documentation. Please let me know if this works for you.
>
> Thanks!

Great! I'll give it a shot again!

@dfm I have mostly completed my review, just was trying to get some examples to run on my local machine.
