### 2024-05-06 #6573 @michael-petersen
Hi,

@iniyannatarajan Thanks for the thoughts on MacOS install, it seems like that is still a work in progress, which is fine. Maybe worth noting in the installation documentation?

An Ubuntu installation is very smooth, but I agree with @kiranshila that the tutorial currently refers to some local paths (e.g. the tutorial [here](https://iniyannatarajan.github.io/Anime.jl/v0.3/examples/createdataset/)) and I haven't been able to run any of the examples locally. @iniyannatarajan I'll open an issue to this effect, since it looks like you might be able to hack this with `relativepath = joinpath(pathof(Anime),"../../")`, but I'm not sure that is behaving as intended.
