### 2024-05-09 4307c32
Add note regarding Mac installation issues to docs
