+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"components/#Components","page":"Components","title":"Components","text":"","category":"section"},{"location":"components/","page":"Components","title":"Components","text":"Anime can be run in two modes: imported as a module with callable functions or as a single pipeline that accepts input settings in YAML and CSV files and runs without human intervention. The following diagram gives an overview of the design.","category":"page"},{"location":"components/","page":"Components","title":"Components","text":"(Image: Design overview)","category":"page"},{"location":"components/","page":"Components","title":"Components","text":"Anime loads the metadata of the VLBI observation and the observed data into a Structure-of-Arrays (SoA) in-memory representation with which all the functions that compute instrument models are expected to interface. The generated models (also referred to as gain tables) are written out to an HDF5 file. The functionality within the dotted boxes are not considered essential to the generation of instrument models, but are provided for completion. These include","category":"page"},{"location":"components/","page":"Components","title":"Components","text":"generation of an MS from scratch,\ncomputing and writing source coherency to MS, and\napplying instrument models to data and writing them back to the MS.","category":"page"},{"location":"components/","page":"Components","title":"Components","text":"In modular mode, any calibration or imaging software can call any function in the box \"Compute instrument model\" and process the gain tables as they see fit. In pipeline mode, the entire control flow is automated by using a YAML file as shown below.","category":"page"},{"location":"components/#Input-files","page":"Components","title":"Input files","text":"","category":"section"},{"location":"components/#Observation-parameters","page":"Components","title":"Observation parameters","text":"","category":"section"},{"location":"components/","page":"Components","title":"Components","text":"Functions such as createmsfromconfig require the user to input various arguments that define the parameters of observation. For ease of use, these parameters can be collected in a YAML file which can be passed directly to Anime in pipeline mode.","category":"page"},{"location":"components/","page":"Components","title":"Components","text":"####################################################\n# sample config file for AnimeDriver.jl\n####################################################\n# settings to generate a new ms\n\nmode: \"manual\"                                       # \"uvfits\" and \"manual\" are the only allowed values\nmsname: \"eht.ms\"                                  # name of output ms \ntelescopename: \"VLBA\"                                # preset telescope known to casa (not important since CASA is not used for predict; just an argument required for setconfig)\nstations: \"inputs/eht_2017.stations\"     # CSV file with station parameters\ncasaanttemplate: \"inputs/antenna_table.template\"                 # Input CASA ANTENNA table to use as template\nvex: \"inputs/vexfiles/e17e11.vex\"    # name of the VEX schedule to use for observing\nuvfits: \"inputs/vfitsfiles/hops_lo_3601_M87+zbl-dtcal_selfcal.uvfits\" # name of uvfits file to convert to ms\nautocorr: false                                      # toggle whether to include autocorrelations\nfeed: \"perfect R L\"                                  # \"perfect R L\" or \"perfect X Y\"\nshadowlimit: 1e-6                                    # maximum fraction of geome
[…]
