### 2024-05-09 b48ed54
Update install note for Mac
