### 2024-05-09 #6573 @iniyannatarajan
Hi Michael,

Thanks for your comments!

I have added a note regarding MacOS in the installation section of the documentation [here](https://iniyannatarajan.github.io/Anime.jl/dev/install/). Downgrading `casatools` to version 6.5.1.23 in `CondaPkg.toml` could be a solution for MacOS as Lindy Blackburn has mentioned [here](https://github.com/iniyannatarajan/Anime.jl/issues/26#issuecomment-2089253699), but I have currently left it at version 6.5.5.21 which has been used for all of the testing so far.

I have updated the instructions for downloading the source code and generating all the docs & tutorials locally in the [tutorial section](https://iniyannatarajan.github.io/Anime.jl/dev/examples/createdataset/) of the documentation. I have also removed the relative path loading from the documentation to avoid confusion (it is still internally used), so that anyone with `Anime` installed in their local Julia environment can just type `using Anime` and step through the tutorials.

Please note that I have refactored the code since v0.3 or v0.4 and the latest code in the main branch (and the dev docs) reflects the changes mentioned above. Let me know if the above changes work for you.
