### 2024-05-10 2bf2bac
CompatHelper: bump compat for CairoMakie to 0.12, (keep existing compat)
