+++ Project.toml
@@ -37,7 +37,7 @@ YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 ArgParse = "1.1"
 CFITSIO_jll = "4.3"
 CSV = "0.10"
-CairoMakie = "0.11"
+CairoMakie = "0.11, 0.12"
 Casacore = "0.3"
 ColorSchemes = "3.21"
 CondaPkg = "0.2"
