### 2024-05-10 5619f7d
Add instructions to install dev version
