### 2024-05-13 #6573 @kiranshila
@iniyannatarajan I'm still having trouble installing on the main branch (on Arch linux). I tried from a cleared Julia and pip cache with the same results. Looks like now I'm getting a numpy version problem? I've attached the log.

```julia
ERROR: The following 1 direct dependency failed to precompile:

Anime [628eb91b-9d1c-424a-81d5-cd7fa5d7294a]

Failed to precompile Anime [628eb91b-9d1c-424a-81d5-cd7fa5d7294a] to "/home/kiran/.julia/compiled/v1.10/Anime/jl_snnWTH".
    CondaPkg Found dependencies: /home/kiran/.julia/packages/Anime/d9tOh/CondaPkg.toml
    CondaPkg Found dependencies: /home/kiran/.julia/packages/PythonCall/wXfah/CondaPkg.toml
    CondaPkg Resolving changes
             + casadata (pip)
             + casatasks (pip)
             + casatools (pip)
             + libstdcxx-ng
             + openssl
             + pip
             + python
    CondaPkg Creating environment
             │ /home/kiran/.julia/artifacts/7973f2c7725e2d0eef7a95159454c4145f0945a2/bin/micromamba
             │ -r /home/kiran/.julia/scratchspaces/0b3b1443-0f03-428d-bdfb-f27f9c1191ea/root
             │ create
             │ -y
             │ -p /home/kiran/.julia/environments/v1.10/.CondaPkg/env
             │ --override-channels
             │ --no-channel-priority
             │ libstdcxx-ng[version='>=3.4,<13.0']
             │ openssl[version='=3.0.2']
             │ pip[version='>=22.0.0']
             │ python[version='>=3.7,<4',channel='conda-forge',build='*cpython*']
             │ python[version='=3.8']
             └ -c conda-forge
conda-forge/linux-64                                        Using cache
conda-forge/noarch                                          Using cache

Transaction

  Prefix: /home/kiran/.julia/environments/v1.10/.CondaPkg/env

  Updating specs:

   - libstdcxx-ng[version='>=3.4,<13.0']
   - openssl===3.0.2
   - pip[version='>=22.0.0']
   - conda-forge::python[version='>=3.7,<4',build=*cpython*]
   - python===3.8

  Package              Version  Build               Channel           Size
────────────────────────────────────────────────────────────────────────────
  Install:
────────────────────────────────────────────────────────────────────────────

  + libstdcxx-ng        12.3.0  hc0a3c3a_7          conda-forge     Cached
  + _libgcc_mutex          0.1  conda_forge         conda-forge     Cached
  + ca-certificates   2024.2.2  hbcca054_0          conda-forge     Cached
  + ld_impl_linux-64      2.40  h55db66e_0          conda-forge     Cached
  + libgomp             13.2.0  h77fa898_7          conda-forge     Cached
  + _openmp_mutex          4.5  2_gnu               conda-forge     Cached
  + libgcc-ng           13.2.0  h77fa898_7          conda-forge     Cached
  + xz                   5.2.6  h166bdaf_0          conda-forge     Cached
  + ncurses                6.5  h59595ed_0          conda-forge     Cached
  + libzlib             1.2.13  hd590300_5          conda-forge     Cached
  + libuuid             2.38.1  h0b41bf4_0          conda-forge     Cached
  + libnsl               2.0.1  hd590300_0          conda-forge     Cached
  + libffi               3.4.2  h7f98852_5          conda-forge     Cached
  + bzip2                1.0.8  hd590300_5          conda-forge     Cached
  + openssl              3.0.2  h166bdaf_1          conda-forge     Cached
  + readline               8.2  h8228510_1          conda-forge     Cached
  + tk                  8.6.13  noxft_h4845f30_101  conda-forge     Cached
  + libsqlite           3.45.3  h2797004_0          conda-forge     Cached
  + sqlite              3.45.3  h2c6b66d_0          conda-forge     Cached
  + python              3.8.13  ha86cf86_0_cpython  conda-forge     Cached
  + wheel               0.43.0  pyhd8ed1ab_1        conda-forge     Cached
  + setuptools          69.5.1  pyhd8ed1ab_0        conda-forge     Cached
  + pip                   24.0  pyhd8ed1ab_0        conda-forge     Cached

  Summary:

  Install: 23 packages

  Total download: 0 B

────────────────────────────────────────────────────────────────────────────

Transaction starting
Linking libstdcxx-ng-12.3.0-hc0a3c3a_7
Linking _libgcc_mutex-0.1-conda_forge
Linking ca-certificates-2024.2.2-hbcca054_0
Linking ld_impl_linux-64-2.40-h55db66e_0
Linking libgomp-13.2.0-h77fa898_7
Linking _openmp_mutex-4.5-2_gnu
Linking libgcc-ng-13.2.0-h77fa898_7
Linking xz-5.2.6-h166bdaf_0
Linking ncurses-6.5-h59595ed_0
Linking libzlib-1.2.13-hd590300_5
Linking libuuid-2.38.1-h0b41bf4_0
Linking libnsl-2.0.1-hd590300_0
Linking libffi-3.4.2-h7f98852_5
Linking bzip2-1.0.8-hd590300_5
Linking openssl-3.0.2-h166bdaf_1
Linking readline-8.2-h8228510_1
Linking tk-8.6.13-noxft_h4845f30_101
Linking libsqlite-3.45.3-h2797004_0
Linking sqlite-3.45.3-h2c6b66d_0
Linking python-3.8.13-ha86cf86_0_cpython
Linking wheel-0.43.0-pyhd8ed1ab_1
Linking setuptools-69.5.1-pyhd8ed1ab_0
Linking pip-24.0-pyhd8ed1ab_0

Transaction finished

To activate this environment, use:

    micromamba activate /home/kiran/.julia/environments/v1.10/.CondaPkg/env

Or to execute a single command in this environment, use:

    micromamba run -p /home/kiran/.julia/environments/v1.10/.CondaPkg/env mycommand

    CondaPkg Installing Pip packages
             │ /home/kiran/.julia/environments/v1.10/.CondaPkg/env/bin/pip
             │ install
             │ casadata
             │ casatasks ==6.5.5.21
             └ casatools ==6.5.5.21
Collecting casadata
  Downloading casadata-2024.1.15-py3-none-any.whl.metadata (430 bytes)
Collecting casatasks==6.5.5.21
  Downloading casatasks-6.5.5.21-py3-none-any.whl.metadata (685 bytes)
Collecting casatools==6.5.5.21
  Downloading casatools-6.5.5.21-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (1.1 kB)
Collecting matplotlib (from casatasks==6.5.5.21)
  Downloading matplotlib-3.7.5-cp38-cp38-manylinux_2_12_x86_64.manylinux2010_x86_64.whl.metadata (5.7 kB)
Requirement already satisfied: scipy in ./.local/lib/python3.8/site-packages (from casatasks==6.5.5.21) (1.4.1)
Collecting certifi (from casatasks==6.5.5.21)
  Downloading certifi-2024.2.2-py3-none-any.whl.metadata (2.2 kB)
Collecting numpy (from casatools==6.5.5.21)
  Downloading numpy-1.24.4-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.6 kB)
Collecting contourpy>=1.0.1 (from matplotlib->casatasks==6.5.5.21)
  Downloading contourpy-1.1.1-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (5.9 kB)
Collecting cycler>=0.10 (from matplotlib->casatasks==6.5.5.21)
  Downloading cycler-0.12.1-py3-none-any.whl.metadata (3.8 kB)
Collecting fonttools>=4.22.0 (from matplotlib->casatasks==6.5.5.21)
  Downloading fonttools-4.51.0-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl.metadata (159 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 159.5/159.5 kB 5.6 MB/s eta 0:00:00
Collecting kiwisolver>=1.0.1 (from matplotlib->casatasks==6.5.5.21)
  Downloading kiwisolver-1.4.5-cp38-cp38-manylinux_2_5_x86_64.manylinux1_x86_64.whl.metadata (6.4 kB)
Collecting packaging>=20.0 (from matplotlib->casatasks==6.5.5.21)
  Downloading packaging-24.0-py3-none-any.whl.metadata (3.2 kB)
Requirement already satisfied: pillow>=6.2.0 in ./.local/lib/python3.8/site-packages (from matplotlib->casatasks==6.5.5.21) (7.1.1)
Collecting pyparsing>=2.3.1 (from matplotlib->casatasks==6.5.5.21)
  Downloading pyparsing-3.1.2-py3-none-any.whl.metadata (5.1 kB)
Collecting python-dateutil>=2.7 (from matplotlib->casatasks==6.5.5.21)
  Downloading python_dateutil-2.9.0.post0-py2.py3-none-any.whl.metadata (8.4 kB)
Collecting importlib-resources>=3.2.0 (from matplotlib->casatasks==6.5.5.21)
  Downloading importlib_resources-6.4.0-py3-none-any.whl.metadata (3.9 kB)
Collecting zipp>=3.1.0 (from importlib-resources>=3.2.0->matplotlib->casatasks==6.5.5.21)
  Downloading zipp-3.18.1-py3-none-any.whl.metadata (3.5 kB)
Collecting six>=1.5 (from python-dateutil>=2.7->matplotlib->casatasks==6.5.5.21)
  Downloading six-1.16.0-py2.py3-none-any.whl.metadata (1.8 kB)
Downloading casatasks-6.5.5.21-py3-none-any.whl (1.8 MB)
Downloading casatools-6.5.5.21-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (133.6 MB)
Downloading casadata-2024.1.15-py3-none-any.whl (349.3 MB)
Downloading certifi-2024.2.2-py3-none-any.whl (163 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 163.8/163.8 kB 15.3 MB/s eta 0:00:00
Downloading matplotlib-3.7.5-cp38-cp38-manylinux_2_12_x86_64.manylinux2010_x86_64.whl (9.2 MB)
Downloading numpy-1.24.4-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (17.3 MB)
Downloading contourpy-1.1.1-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (301 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 301.1/301.1 kB 31.0 MB/s eta 0:00:00
Downloading cycler-0.12.1-py3-none-any.whl (8.3 kB)
Downloading fonttools-4.51.0-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (4.7 MB)
Downloading importlib_resources-6.4.0-py3-none-any.whl (38 kB)
Downloading kiwisolver-1.4.5-cp38-cp38-manylinux_2_5_x86_64.manylinux1_x86_64.whl (1.2 MB)
Downloading packaging-24.0-py3-none-any.whl (53 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 53.5/53.5 kB 5.7 MB/s eta 0:00:00
Downloading pyparsing-3.1.2-py3-none-any.whl (103 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 103.2/103.2 kB 11.3 MB/s eta 0:00:00
Downloading python_dateutil-2.9.0.post0-py2.py3-none-any.whl (229 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 229.9/229.9 kB 18.9 MB/s eta 0:00:00
Downloading six-1.16.0-py2.py3-none-any.whl (11 kB)
Downloading zipp-3.18.1-py3-none-any.whl (8.2 kB)
Installing collected packages: casadata, zipp, six, pyparsing, packaging, numpy, kiwisolver, fonttools, cycler, certifi, python-dateutil, importlib-resources, contourpy, casatools, matplotlib, casatasks
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
librosa 0.7.2 requires decorator>=3.0.0, which is not installed.
pandas 1.0.3 requires pytz>=2017.2, which is not installed.
tb-nightly 2.3.0a20200424 requires markdown>=2.6.8, which is not installed.
tb-nightly 2.3.0a20200424 requires requests<3,>=2.21.0, which is not installed.
tensorboard 2.2.1 requires markdown>=2.6.8, which is not installed.
tensorboard 2.2.1 requires requests<3,>=2.21.0, which is not installed.
tf-nightly 2.1.0 requires wrapt>=1.11.1, which is not installed.
Successfully installed casadata-2024.1.15 casatasks-6.5.5.21 casatools-6.5.5.21 certifi-2024.2.2 contourpy-1.1.1 cycler-0.12.1 fonttools-4.51.0 importlib-resources-6.4.0 kiwisolver-1.4.5 matplotlib-3.7.5 numpy-1.24.4 packaging-24.0 pyparsing-3.1.2 python-dateutil-2.9.0.post0 six-1.16.0 zipp-3.18.1
ERROR: LoadError: Python: ImportError: cannot import name 'int' from 'numpy' (/home/kiran/.julia/environments/v1.10/.CondaPkg/env/lib/python3.8/site-packages/numpy/__init__.py)
Python stacktrace:
 [1] <module>
   @ ~/.local/lib/python3.8/site-packages/scipy/special/orthogonal.py:81
 [2] <module>
   @ ~/.local/lib/python3.8/site-packages/scipy/special/_basic.py:19
 [3] <module>
   @ ~/.local/lib/python3.8/site-packages/scipy/special/__init__.py:637
 [4] <module>
   @ ~/.julia/environments/v1.10/.CondaPkg/env/lib/python3.8/site-packages/casatasks/private/JPLephem_reader2.py:21
 [5] <module>
   @ ~/.julia/environments/v1.10/.CondaPkg/env/lib/python3.8/site-packages/casatasks/private/task_fixplanets.py:10
 [6] <module>
   @ ~/.julia/environments/v1.10/.CondaPkg/env/lib/python3.8/site-packages/casatasks/fixplanets.py:9
 [7] <module>
   @ ~/.julia/environments/v1.10/.CondaPkg/env/lib/python3.8/site-packages/casatasks/__init__.py:178
Stacktrace:
 [1] pythrow()
   @ PythonCall ~/.julia/packages/PythonCall/wXfah/src/err.jl:94
 [2] errcheck
   @ ~/.julia/packages/PythonCall/wXfah/src/err.jl:10 [inlined]
 [3] pyimport(m::String)
   @ PythonCall ~/.julia/packages/PythonCall/wXfah/src/concrete/import.jl:11
 [4] pyimport(::Pair{String, String})
   @ PythonCall ~/.julia/packages/PythonCall/wXfah/src/concrete/import.jl:12
 [5] top-level scope
   @ ~/.julia/packages/Anime/d9tOh/src/Anime.jl:22
 [6] include
   @ ./Base.jl:495 [inlined]
 [7] include_package_for_output(pkg::Base.PkgId, input::String, depot_path::Vector{String}, dl_load_path::Vector{String}, load_path::Vector{String}, concrete_deps::Vector{Pair{Base.PkgId, UInt128}}, source::Nothing)
   @ Base ./loading.jl:2222
 [8] top-level scope
   @ stdin:3
in expression starting at /home/kiran/.julia/packages/Anime/d9tOh/src/Anime.jl:4
in expression starting at stdin:
Stacktrace:
  [1] pkgerror(msg::String)
    @ Pkg.Types ~/.julia/juliaup/julia-1.10.3+0.x64.linux.gnu/share/julia/stdlib/v1.10/Pkg/src/Types.jl:70
  [2] precompile(ctx::Pkg.Types.Context, pkgs::Vector{…}; internal_call::Bool, strict::Bool, warn_loaded::Bool, already_instantiated::Bool, timing::Bool, _from_loading::Bool, kwargs::@Kwargs{…})
    @ Pkg.API ~/.julia/juliaup/julia-1.10.3+0.x64.linux.gnu/share/julia/stdlib/v1.10/Pkg/src/API.jl:1659
  [3] precompile(pkgs::Vector{Pkg.Types.PackageSpec}; io::Base.TTY, kwargs::@Kwargs{_from_loading::Bool})
    @ Pkg.API ~/.julia/juliaup/julia-1.10.3+0.x64.linux.gnu/share/julia/stdlib/v1.10/Pkg/src/API.jl:159
  [4] precompile
    @ ~/.julia/juliaup/julia-1.10.3+0.x64.linux.gnu/share/julia/stdlib/v1.10/Pkg/src/API.jl:147 [inlined]
  [5] #precompile#114
    @ ~/.julia/juliaup/julia-1.10.3+0.x64.linux.gnu/share/julia/stdlib/v1.10/Pkg/src/API.jl:146 [inlined]
  [6] #invokelatest#2
    @ ./essentials.jl:894 [inlined]
  [7] invokelatest
    @ ./essentials.jl:889 [inlined]
  [8] _require(pkg::Base.PkgId, env::String)
    @ Base ./loading.jl:1963
  [9] __require_prelocked(uuidkey::Base.PkgId, env::String)
    @ Base ./loading.jl:1812
 [10] #invoke_in_world#3
    @ ./essentials.jl:926 [inlined]
 [11] invoke_in_world
    @ ./essentials.jl:923 [inlined]
 [12] _require_prelocked(uuidkey::Base.PkgId, env::String)
    @ Base ./loading.jl:1803
 [13] macro expansion
    @ ./loading.jl:1790 [inlined]
 [14] macro expansion
    @ ./lock.jl:267 [inlined]
 [15] __require(into::Module, mod::Symbol)
    @ Base ./loading.jl:1753
 [16] #invoke_in_world#3
    @ ./essentials.jl:926 [inlined]
 [17] invoke_in_world
    @ ./essentials.jl:923 [inlined]
 [18] require(into::Module, mod::Symbol)
    @ Base ./loading.jl:1746
Some type information was truncated. Use `show(err)` to see complete types.
```
