### 2024-05-13 #6573 @kiranshila
It seems the scipy version needs to be bumped (as a transitive dep of casatasks) as numpy 1.24 doesn't have a `np.int` type, which scipy refers to.
