+++ CondaPkg.toml
@@ -6,3 +6,4 @@ openssl = "=3.0.2"
 casadata = ""
 casatools = "==6.5.5.21"
 casatasks = "==6.5.5.21"
+numpy = "==1.23.5"
