### 2024-05-14 #6573 @kiranshila
That did the trick! Classic python versioning, I blame CASA. I'm surprised this didn't come up in tests, I guess things were using an older cached version? Like, you'd think casatools would have noticed.

I was able to run the tutorials manually by cloning the repo and setting `relativepath` to the cloned dir! Everything is now working as expected. Great job on the docs updates, everything looks good!

Maybe a last little note (unrelated to the review), this is still a relatively complex software stack with c libraries, julia, python, and all sorts of crazy interlanguage things happening. It might be in your interest to explore some deterministic packaging to create environments for your users with the likes of Nix, Guix, or Spack. I've been exploring this with our telescope project with decent success. Maybe something to check out if you get bored :)

@dfm I've completed my review and recommend this paper for publication
