### 2024-05-14 #6573 @iniyannatarajan
Thanks @kiranshila for your comments!

The logs of previous builds on ubuntu-latest show that numpy 1.24.4 and scipy 1.10.1 were being installed and the tests had run without issues (same on my local Ubuntu machine too), which is puzzling to me.

I will check out the packaging tools you have recommended. The idea is to ultimately eliminate the reliance on external software and python dependencies. Only MS creation from scratch uses python casatools/casatasks directly. To avoid this, Casacore.jl needs to support the simulator tool. If I get time I will see if I can add this to Casacore.jl.
