### 2024-05-14 #6573 @michael-petersen
Hi @iniyannatarajan ,

I'm able to run _most_ of the tutorial now on Ubuntu, no problem -- thanks for the updates!

However, I'm still crashing out of the first tutorial (['Creating Data Sets'](https://iniyannatarajan.github.io/Anime.jl/dev/examples/createdataset/)) with an opaque segfault that I'm not sure how to diagnose. @kiranshila , you were able to run this tutorial fine?
<details><summary>Error log</summary>

```
julia> createmsfromuvfits(uvfits, msname, mode)

signal (11): Segmentation fault
in expression starting at REPL[41]:1
_PyVectorcall_Function at /usr/local/src/conda/python-3.8.13/Include/cpython/abstract.h:83 [inlined]
PyObject_Call at /usr/local/src/conda/python-3.8.13/Objects/call.c:227
unknown function (ip: 0x7f01cc29c06f)
Allocations: 260147087 (Pool: 260072308; Big: 74779); GC: 73
Segmentation fault
```

With the same sort of behaviour for the other data sets examples.

I'm guessing it is some sort of Python issue under the hood, but for the purposes of this test, I'm just blindly following the documentation.

The docs for ['Compute coherency matrix'](https://iniyannatarajan.github.io/Anime.jl/dev/examples/computecoherency/) still need the path fixes in the dev version.

I'm very happy with the other changes and the text of the paper is great. Just want to see the tutorial to some sort of conclusion and then am happy to sign off!
