### 2024-05-15 #6573 @iniyannatarajan
Hi @michael-petersen, thanks for catching this!

I have updated the documentation with the path fixes and additional clarification, but I too get the same error as you when I create a temporary environment in Julia REPL, install Anime from GitHub, and call `createmsfromuvfits`.

It does not happen if I do something like `include("/path/to/Anime.jl/src/Anime.jl")` in this temporary environment (which replaces module Anime with a warning) and then call the same function; or if I do `importuvfits = pyimport("casatasks" => "importuvfits")` and then call `importuvfits` directly.

I will add an instruction to this effect in the tutorial for creating data sets for now and look into this issue.
