### 2024-05-16 #6573 @michael-petersen
To conclude my review, just a few more compliments on the paper; I really liked the control flow figure, and appreciated the inclusion of a 'Related Software' section, as someone not familiar with the radio astronomy Julia offerings. Great work!

@dfm , I have completed my review and recommend the paper for publication.
