### 2024-05-16 #6573 @iniyannatarajan
Hi @michael-petersen , thanks for your comments!

I have added the instruction to the docs [here](https://iniyannatarajan.github.io/Anime.jl/dev/examples/createdataset/#Creating-data-sets). I will make a new release with the updates made since version 0.4.0 (plus additional minor updates) in the next week or so.

Getting rid of the python dependencies is indeed a high priority item now without losing the ability to handle CASA data sets. New data schemas are being developed by a few different VLBI/radio astronomy groups, for which I (and a few others) are interested in developing native support in Julia. Ultimately, we would like to move to these as the primary formats for storing data and metadata.
