### 2024-05-16 #6573 @iniyannatarajan
@kiranshila Thanks! I am indeed about to replace AATM with AM for atmospheric modelling and have already built a binary wrapper [here](https://github.com/JuliaBinaryWrappers/AM_jll.jl) for it. This will replace AATM in future releases.

I will try to build a binary wrapper for `WSClean`. The idea is to reproduce its functionality natively in Julia (similar to packages such as `Comrade.jl`) but that is slightly longer term for now.
