+++ paper/paper.md
@@ -38,7 +38,7 @@ bibliography: paper.bib
 [^1]: https://julialang.org
 
 # Statement of need
-The Event Horizon Telescope (EHT) [@M87PaperII] is a global mm-VLBI network that produced the first event-horizon-scale images of supermassive black holes at the centres of M87 [@M87PaperI] and the Milky Way [@SgrAPaperI] galaxies. Existing radio telescopes are regularly added to the network, while new geographic locations are also being considered for future stations. A physics-based approach to model atmospheric and instrumental characteristics is essential to optimize the design of new instruments and to characterize existing ones. Such a "forward modelling" approach aids in developing new algorithms for processing data from sparse, heterogeneous VLBI arrays such as the EHT and future next-generation upgrades to the EHT (ngEHT) [@Doeleman2023].
+The Event Horizon Telescope [EHT, @M87PaperII] is a global mm-VLBI network that produced the first event-horizon-scale images of supermassive black holes at the centres of M87 [@M87PaperI] and the Milky Way [@SgrAPaperI] galaxies. Existing radio telescopes are regularly added to the network, while new geographic locations are also being considered for future stations. A physics-based approach to model atmospheric and instrumental characteristics is essential to optimize the design of new instruments and to characterize existing ones. Such a "forward modelling" approach aids in developing new algorithms for processing data from sparse, heterogeneous VLBI arrays such as the EHT and future next-generation upgrades to the EHT [ngEHT, @Doeleman2023].
 
 In radio astronomy parlance, "calibration" refers to the act of removing atmospheric and instrumental effects from the data while "imaging" refers to the act of reconstructing the sky brightness distribution from the interferometric data. While some calibration parameters may be constrained a priori, accurate calibration parameters are often fitted to the data directly through a process of "self-calibration" [@Pearson1984], which may be performed either iteratively or simultaneously with imaging depending on software capability. Realistic instrument models thus enable both statistically efficient and accurate calibration, leading to a more faithful reconstruction of sky brightness from the data.
 
@@ -50,7 +50,7 @@ The metadata for generating instrument models are loaded in-memory from various 
 
 ![Components and control flow of a typical modelling run.](anime-components.png)
 
-A generic instrument model may vary along station, time, frequency and polarization axes. `Anime` includes models for the troposphere which significantly affects signal propagation at mm-wavelengths (86 GHz and above). It also models the instrumental contribution to signal polarization, telescope tracking offsets, bandpass effects and receiver electronic gains, alongside the noise contributions from the atmosphere and receiver electronics. All time-variable effects are modelled using Gaussian Processes (GPs) [@GPML2006], with the hyperparameters chosen to statistically match the empirically measured temporal correlation structure of the quantity being modelled.
+A generic instrument model may vary along station, time, frequency and polarization axes. `Anime` includes models for the troposphere which significantly affects signal propagation at mm-wavelengths (86 GHz and above). It also models the instrumental contribution to signal polarization, telescope tracking offsets, bandpass effects and receiver electronic gains, alongside the noise contributions from the atmosphere and receiver electronics. All time-variable effects are modelled using Gaussian Processes [GPs, @GPML2006], with the hyperparameters chosen to statistically match the empirically measured temporal correlation structure of the quantity being modelled.
 
 Synthetic data generation capabilities are built into `Anime`, with support for popular VLBI data storage formats such as UVFITS and Measurement Sets (MS)[^2]. The generated instrument models are applied to the uncorrupted data using the Radio Interferometer Measurement Equation (RIME) that describes the full polarization state of the signal with the 2x2 _Jones matrix_ formalism [@OMS2011]:
 
