+++ paper/paper.bib
@@ -23,7 +23,7 @@ journal = {Journal of Open Source Software} }
 @ARTICLE{Chael2018,
        author = {{Chael}, Andrew A. and {Johnson}, Michael D. and {Bouman}, Katherine L. and {Blackburn}, Lindy L. and {Akiyama}, Kazunori and {Narayan}, Ramesh},
         title = "{Interferometric Imaging Directly with Closure Phases and Closure Amplitudes}",
-      journal = {Astrophysical Journal},
+      journal = {The Astrophysical Journal},
      keywords = {accretion, accretion disks, black hole physics, Galaxy: center, techniques: high angular resolution, techniques: image processing, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
          year = 2018,
         month = apr,
@@ -54,7 +54,7 @@ archivePrefix = {arXiv},
 @ARTICLE{M87PaperI,
    author = {{Event Horizon Telescope Collaboration}},
     title = "{First M87 Event Horizon Telescope Results. I. The Shadow of the Supermassive Black Hole}",
-  journal = {Astrophysical Journal, Letters},
+  journal = {The Astrophysical Journal Letters},
  keywords = {accretion, accretion disks, black hole physics, galaxies: active, galaxies: individual: M87, galaxies: jets, gravitation },
      year = 2019,
     month = apr,
@@ -69,7 +69,7 @@ archivePrefix = {arXiv},
 @ARTICLE{M87PaperII,
    author = {{Event Horizon Telescope Collaboration}},
     title = "{First M87 Event Horizon Telescope Results. II. Array and Instrumentation}",
-  journal = {Astrophysical Journal, Letters},
+  journal = {The Astrophysical Journal Letters},
  keywords = {black hole physics, galaxies: individual: M87, Galaxy: center, gravitational lensing: strong, instrumentation: interferometers, techniques: high angular resolution, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies, Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
      year = 2019,
     month = apr,
@@ -117,7 +117,7 @@ archivePrefix = {arXiv},
 @ARTICLE{SgrAPaperI,
        author = {{Event Horizon Telescope Collaboration}},
         title = "{First Sagittarius A* Event Horizon Telescope Results. I. The Shadow of the Supermassive Black Hole in the Center of the Milky Way}",
-      journal = {Astrophysical Journal, Letters},
+      journal = {The Astrophysical Journal Letters},
      keywords = {Black holes, Kerr black holes, Rotating black holes, Heterodyne interferometry, Galactic center, 162, 886, 1406, 726, 565},
          year = 2022,
         month = may,
@@ -152,7 +152,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Natarajan2022,
        author = {{Natarajan}, Iniyan and {Deane}, Roger and {Mart{\'\i}-Vidal}, Iv{\'a}n and {Roelofs}, Freek and {Janssen}, Michael and {Wielgus}, Maciek and {Blackburn}, Lindy and {Blecher}, Tariq and {Perkins}, Simon and {Smirnov}, Oleg and {Davelaar}, Jordy and {Moscibrodzka}, Monika and {Chael}, Andrew and {Bouman}, Katherine L. and {Kim}, Jae-Young and {Bernardi}, Gianni and {van Bemmel}, Ilse and {Falcke}, Heino and {{\"O}zel}, Feryal and {Psaltis}, Dimitrios},
[…]
