### 2024-05-21 cc2bf5a
Update sample code in paper
