### 2024-05-21 #6573 @iniyannatarajan
@dfm Thanks for the PR. I have made some minor updates to the paper draft and created a new release and DOI for the software.

Version: v0.5.0
DOI: 10.5281/zenodo.11238336
