+++ paper/paper.bib
@@ -226,7 +226,7 @@ archivePrefix = {arXiv},
 series = {Adaptive computation and machine learning},
 publisher = {MIT Press},
 isbn = {9786612097966},
-doi = {https://doi.org/10.7551/mitpress/3206.001.0001},
+doi = {10.7551/mitpress/3206.001.0001},
 year = {2006},
 title = {Gaussian processes for machine learning},
 language = {eng},
